users-service/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0

pydantic==2.12.4
pydantic-settings==2.11.0
python-dotenv==1.2.1
pydantic[email]==2.12.4
SQLAlchemy==2.0.44
greenlet==3.2.4
asyncpg==0.30.0
alembic==1.14.0

python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
anyio==4.11.0
click==8.3.0
typing_extensions==4.15.0


